refactor: back card fix
