+++ src/component/result/IDCard/front/right-box.tsx
@@ -34,31 +34,6 @@ const TopWrapper = styled.div`
   height: 120px;
 `;
 
-const FrontTextWrapper = styled.div`
-  height: 120px;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  justify-content: space-between;
-  padding: 12px 10px;
-
-  background: ${(props) => props.theme.colors.sub2};
-  color: ${(props) => props.theme.colors.main2};
-
-  font-family: 'GongGothicMedium';
-  font-weight: 500;
-  font-size: 10px;
-  line-height: 12px;
-
-  border-radius: 16px;
-  div {
-    border-bottom: 0.1px solid #343232;
-    width: 100%;
-    display: flex;
-    justify-content: space-between;
-  }
-`;
-
 const BottomWrapper = styled.div``;
 
 const Desc = styled.div`
@@ -66,7 +41,7 @@ const Desc = styled.div`
   line-height: 10px;
   font-size: 10px;
 
-  transform: translate(0%, 80%) scale(0.75);
+  transform: translate(0%, 55%) scale(0.75);
 
   display: inline-block;
   text-align: right;
